backend/requirements.txt
fastapi
uvicorn
twilio
python-dotenv
elevenlabs
starlette
python-multipart
pydantic
sqlmodel
stripe
python-jose[cryptography]
passlib[bcrypt]
bcrypt
